Mostly Folder Controller work and bug fixes

+++ BLL/Services/FolderService.cs
@@ -9,6 +9,8 @@ using AutoMapper;
 using System.Threading.Tasks;
 using System.Linq;
 using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Authorization;
+using System.Data.SqlTypes;
 
 namespace BLL.Services
 {
@@ -16,15 +18,54 @@ namespace BLL.Services
     {
         private IUnitOfWork _uow;
         private IMapper _mapper;
-        private UserManager<User> _userManager;
 
-        public FolderService( IUnitOfWork uow, IMapper mapper,
-                UserManager<User> userManager)
+
+        public FolderService( IUnitOfWork uow,  IMapper mapper)
         {
             _uow = uow;
             _mapper = mapper;
-            _userManager = userManager;
 
+        }
+        private async Task CloneSubfolders(Guid subfolderId, Guid parentId)
+        {
+            FolderDTO subfolder= await GetFolderById(subfolderId);
+            subfolder.Id = new Guid();
+            subfolder.Owner = null;
+            //FolderDTO newFolder = await CreateAtFolder(subfolder, parentId, subfolder.OwnerId);
+            Folder newFolder = new Folder()
+            {
+                Name = subfolder.Name,
+                Description = subfolder.Description,
+                OwnerId = subfolder.OwnerId,
+                ParentId = subfolder.ParentId
+            };
+            newFolder = await _uow.Folders.Create(newFolder);
+
+
+            if (subfolder.Subfolders != null && subfolder.Subfolders.Any())
+             {
+                foreach (var folder in subfolder.Subfolders)
+                {
+                    await CloneSubfolders(folder.Id, newFolder.Id);
+                }
+             }
+
+            if (subfolder.Files != null && subfolder.Files.Any())
+            {
+                foreach (var file in subfolder.Files)
+                {
+                    await _uow.Files.Create(
+                        new File()
+                        {
+                            Name = file.Name,
+                            OwnerId = file.OwnerId,
+                            Description = file.Description,
+                            URL = file.URL,
+                            FolderId = newFolder.Id
+                        }
+                    );
+                }
+            }
         }
         public async Task<FolderDTO> CopyFolder(Guid folderId)
         {
@@ -32,37 +73,47 @@ namespace BLL.Services
             if (folderCopy == null)
                 throw new ArgumentException("The requested subfolder does not exist");
             folderCopy.Name = "Copy of " + folderCopy.Name;
-
-            folderCopy = await _uow.Folders.Create(folderCopy);
+            folderCopy.Owner = null;
+            var newFolder = new Folder()
+            {
+                Name = "Copy of " + folderCopy.Name,
+                Description = folderCopy.Description,
+                OwnerId = folderCopy.OwnerId,
+                ParentId = folderCopy.ParentId,
+                ShareStatus = folderCopy.ShareStatus
+            };
+            newFolder = await _uow.Folders.Create(newFolder);
+            foreach (var subfolder in folderCopy.Subfolders)
+                await CloneSubfolders(subfolder.Id, newFolder.Id);
 
             return _mapper.Map<FolderDTO>(folderCopy);
 
         }
 
-        public async Task<FolderDTO> CreateAtFolder(FolderDTO folder, Guid folderId, string email)
+        public async Task<FolderDTO> CreateAtFolder(FolderDTO folder, Guid folderId, int userId)
         {
-            User user = await _userManager.FindByEmailAsync(email);
             Folder parent = await _uow.Folders.GetFolderById(folderId);
-            bool fsExists = await _uow.FolderShares.FolderShareExists(folderId, user.Id );
-
+           
             if (parent == null)
                 throw new ArgumentException("The parent folder does not exist");
              
-            if (parent.OwnerId != user.Id && !fsExists)
-                throw new ArgumentException("You do not have access to the folder");
-
             folder.ParentId= folderId;
-            folder.OwnerId = user.Id;
+            folder.OwnerId = userId;
             Folder createdFolder = await _uow.Folders.Create(_mapper.Map<Folder>(folder));
             return _mapper.Map<FolderDTO>(createdFolder);
         }
 
-        public async Task<FolderDTO> CreateAtRoot(FolderDTO folder, string email)
+        public async Task<FolderDTO> CreateAtRoot(FolderDTO folder, int userId)
         {
-            User user = await _userManager.FindByEmailAsync(email);
-            folder.OwnerId = user.Id;
+            folder.OwnerId = userId;
             Folder createdFolder = await _uow.Folders.Create(_mapper.Map<Folder>(folder));
             return _mapper.Map<FolderDTO>(createdFolder);
+        }
+
+        public async Task<FolderDTO> GetFolderById(Guid folderId)
+        {
+            Folder folder =  await _uow.Folders.GetFolderById(folderId);
+            return _mapper.Map<FolderDTO>(folder);
         }
 
         public async Task Delete(Guid folderId)
@@ -71,14 +122,20 @@ namespace BLL.Services
             if (folder == null)
                 throw new ArgumentNullException("The provided folder does not exist");
 
-            foreach (var file in folder.Files)
+            if (folder.Files != null && folder.Files.Any())
             {
-                await _uow.Files.Delete(file.Id);
+                foreach (var file in folder.Files)
+                {
+                    await _uow.Files.Delete(file.Id);
+                }
             }
 
-            foreach ( var subfolder in folder.Subfolders)
+            if (folder.Subfolders != null && folder.Subfolders.Any())
             {
-                await Delete(subfolder.Id);
+                foreach (var subfolder in folder.Subfolders)
+                {
+                    await Delete(subfolder.Id);
+                }
             }
             await _uow.Folders.Delete(folderId);
         }
@@ -87,7 +144,8 @@ namespace BLL.Services
         {
             Folder folder = await _uow.Folders.GetFolderById(folderId);
             if (folder == null)
-                throw new ArgumentNullException("The provided folder does not exist");
+                throw new ArgumentNullException("The requested folder does not exist");
+            folder.Subfolders = null;
             return folder.Subfolders.Select(sf => _mapper.Map<FolderDTO>(sf));
         }
